requirements.txt
Django==2.2.3
djangorestframework==3.10.0